refactor: TableGroup 패키지 변경
