+++ src/main/java/kitchenpos/tablegroup/dto/SingleOrderTableCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.dto;
+package kitchenpos.tablegroup.dto;
 
 public class SingleOrderTableCreateRequest {
