+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -16,11 +16,11 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.product.domain.Product;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.TableGroup;
 import kitchenpos.table.dto.OrderTableResponse;
 import kitchenpos.table.dto.TableCreateRequest;
 import kitchenpos.table.dto.TableUpdateEmptyRequest;
 import kitchenpos.table.dto.TableUpdateNumberOfGuestsRequest;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
